bot/pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "BOT for Telegram"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram==3.20.*",
    "environ-config==24.1.*",
]


[dependency-groups]
dev = [
    "dotenv==0.9.*",
    "mypy==1.15.*",
    "pre-commit==4.1.*",
    "pytest==8.3.*",
    "pytest-asyncio==0.25.*",
    "pytest-cov==6.0.*",
    "ruff==0.9.*",
]
[project.scripts]
run_bot = "bot.bot:run"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "ARG", "SIM", "ERA", "PL", "RUF"]
ignore = ["E203", "E501", "RUF003", "RUF001", "ERA001", "F841"]
unfixable = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "double"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["PLR2004", "ARG001", "ERA001"]
